randomizer: Improve the random number generation.
The Mersenne Twister 19937 generator has a huge memory impact and takes long
time to instantiate.  To reduce the amount of time passed on that, make the
`engine` variable in the `randomNumber` function static.

Another issue was that using `std::random::device` generates undeterministic
random numbers.  This is a problem because the Randomizer module is meant to be
used for tests and in that context it is nice to be able to reproduce them.  The
new `genSeed` static function fixes that issue by providing a deterministic seed
and logging it.